+++ frontend/src/queries/nodes/InsightViz/utils.ts
@@ -1,5 +1,5 @@
 import { ActionsNode, BreakdownFilter, EventsNode, InsightQueryNode, TrendsQuery } from '~/queries/schema'
-import { ActionType, ChartDisplayType, IntervalType } from '~/types'
+import { ActionType, ChartDisplayType, InsightModel, IntervalType } from '~/types'
 import { seriesToActionsAndEvents } from '../InsightQuery/utils/queryNodeToFilter'
 import { getEventNamesForAction } from 'lib/utils'
 import {
@@ -8,6 +8,8 @@ import {
     isStickinessQuery,
     isTrendsQuery,
 } from '~/queries/utils'
+import { filtersToQueryNode } from '../InsightQuery/utils/filtersToQueryNode'
+import equal from 'fast-deep-equal'
 
 export const getAllEventNames = (query: InsightQueryNode, allActions: ActionType[]): string[] => {
     const { actions, events } = seriesToActionsAndEvents((query as TrendsQuery).series || [])
@@ -76,4 +78,26 @@ export const getBreakdown = (query: InsightQueryNode): BreakdownFilter | undefin
     } else {
         return undefined
     }
+}
+
+export const getCachedResults = (
+    cachedInsight: Partial<InsightModel> | undefined | null,
+    query: InsightQueryNode
+): Partial<InsightModel> | undefined => {
+    if (
+        !cachedInsight ||
+        cachedInsight.result === null ||
+        cachedInsight.result === undefined ||
+        cachedInsight.filters === undefined
+    ) {
+        return undefined
+    }
+
+    // only set the cached result when the filters match the currently set ones
+    const cachedQueryNode = filtersToQueryNode(cachedInsight.filters)
+    if (!equal(cachedQueryNode, query)) {
+        return undefined
+    }
+
+    return cachedInsight
 }
